Struct tokei::Language

pub struct Language {
    pub blanks: usize,
    pub code: usize,
    pub comments: usize,
    pub files: Vec<PathBuf>,
    pub stats: Vec<Stats>,
    pub lines: usize,
    pub line_comment: Vec<&'static str>,
    pub multi_line: Vec<(&'static str, &'static str)>,
    pub nested: bool,
    pub total_files: usize,
}

Struct representing a single Language.

Fields

blanks

Number of blank lines.

code

Number of lines of code.

comments

Number of comments(both single, and multi-line)

files

A collection of files to be analysed.

stats

A collection of statistics based on the files provide from files

lines

Number of total lines.

line_comment

A collection of single line comments in the language. ie. // in Rust.

multi_line

A collection of tuples representing the start and end of multi line comments. ie. /* comment */ in Rust.

nested

Whether the language supports nested multi line comments or not.

total_files

The total number of files from stats.

Methods

impl Language

fn new(line_comment: Vec<&'static str>, multi_line: Vec<(&'static str, &'static str)>) -> Self

Constructs a new empty Language with the comments provided.

let mut rust = Language::new(vec!["//"], vec![("/*", "*/")]);

fn new_blank() -> Self

Convience constructor for creating a language that has no commenting syntax.

let json = Language::new_blank();

assert_eq!(json.line_comment, vec![]);

fn new_c() -> Self

Convience constructor for creating a language that has the same commenting syntax as C like languages.

let rust = Language::new(vec!["//"], vec![("/*", "*/")]);
let c = Language::new_c();

assert_eq!(rust.line_comment, c.line_comment);
assert_eq!(rust.multi_line, c.multi_line);

fn new_func() -> Self

Convience constructor for creating a language that has the same commenting syntax as ML like languages.

let ocaml = Language::new_multi(vec![("(*", "*)")]);
let coq = Language::new_func();

assert_eq!(ocaml.line_comment, coq.line_comment);
assert_eq!(ocaml.multi_line, coq.multi_line);

fn new_html() -> Self

Convience constructor for creating a language that has the same commenting syntax as HTML like languages.

let xml = Language::new_multi(vec![("<!--", "-->")]);
let html = Language::new_html();

assert_eq!(xml.line_comment, html.line_comment);
assert_eq!(xml.multi_line, html.multi_line);

fn new_hash() -> Self

Convience constructor for creating a language that has the same commenting syntax as Bash.

let bash = Language::new_single(vec!["#"]);
let yaml = Language::new_hash();

assert_eq!(bash.line_comment, yaml.line_comment);
assert_eq!(bash.multi_line, yaml.multi_line);

fn new_multi(multi_line: Vec<(&'static str, &'static str)>) -> Self

Convience constructor for creating a language that only has multi line comments.

let mustache = Language::new_multi(vec![("{{!", "}}")]);

fn new_pro() -> Self

Convience constructor for creating a language that has the same commenting syntax as Prolog.

let prolog = Language::new(vec!["%"], vec![("/*", "*/")]);
let oz = Language::new_pro();

assert_eq!(prolog.line_comment, oz.line_comment);
assert_eq!(prolog.multi_line, oz.multi_line);

fn new_single(line_comment: Vec<&'static str>) -> Self

Convience constructor for creating a language that only has single line comments.

let haskell = Language::new_single(vec!["--"]);

fn is_empty(&self) -> bool

Checks if the language is empty. Empty meaning it doesn't have any statistics.

let rust = Language::new_c();

assert!(rust.is_empty());

fn is_blank(&self) -> bool

Checks if the language doesn't contain any comments.

let json = Language::new_blank();

assert!(json.is_blank());

fn nested(self) -> Self

Specify if the the language supports nested multi line comments.

let mut rust = Language::new(vec!["//"], vec![("/*", "*/")]).nested();
assert!(rust.nested);

fn sort_by(&mut self, category: &str)

Sorts each of the Stats structs contained in the language based on what category is provided panic!'s if given the wrong category.

let mut rust = Language::new_c();
let foo_stats = Stats::new();
let bar_stats = Stats::new();

foo_stats.code += 20;
bar_stats.code += 10;

rust.stats.push(bar_stats.clone());
rust.stats.push(foo_stats.clone());

assert_eq!(rust.stats, vec![bar_stats.clone(), foo_stats.clone()]);

rust.sort_by(CODE);

assert_eq!(rust.stats, vec![foo_stats, bar_stats]);