Struct tokei::Languages

pub struct Languages {
    // some fields omitted
}

A collection of existing languages(List of Languages)

Methods

impl Languages

fn from_cbor<'a, I: Into<&'a [u8]>>(cbor: I) -> Result<Self>

Creates a Languages struct from cbor.

fn from_json<'a, I: Into<&'a [u8]>>(json: I) -> Result<Self>

Creates a Languages struct from json.

let json = r#"{"Rust":{"blanks":5,"code":12,"comments":0,"stats":[{"blanks":5,"code":12,"comments":0,"lines":17,"name":".\\src\\lib\\build.rs"}],"lines":17,"total_files":1}}"#;
let languages = Languages::from_json(&json);
assert!(languages.get_mut(&LanguageType::Rust).unwrap().code == 12)

fn from_yaml<'a, I: Into<&'a [u8]>>(yaml: I) -> Result<Self>

Creates a Languages struct from json.

let yaml = r#"
    ---
    "Rust":
        "blanks": 5
        "code": 12
        "comments": 0
        "lines": 17
        "stats":
            -
            "blanks": 5
            "code": 12
            "comments": 0
            "lines": 17
            "name": ".\\src\\lib\\build.rs"
        "total_files": 1
"#;

let languages = Languages::from_yaml(&yaml);

assert!(languages.get_mut(&LanguageType::Rust).unwrap().code == 12)

fn get_statistics<'a, I>(&mut self, paths: I, ignored: I) where I: Into<Cow<'a, [&'a str]>>

Get statistics from the list of paths provided, and a list ignored keywords to ignore paths containing them.

let languages = Languages::new();
languages.get_statistics(&vec!["."], &vec![".git", "target"]);

println!("{:?}", languages);

fn new() -> Self

Constructs a new, blank Languages.

let languages = Languages::new();

fn to_cbor(&self) -> Result<Vec<u8>, Error>

Converts Languages to CBOR.

extern crate rustc_serialize;
use rustc_serialize::ToHex;

let cbor = "a16452757374a666626c616e6b730564636f64650c68636f6d6d656e74730\
    065737461747381a566626c616e6b730564636f64650c68636f6d6d656e747300656c\
    696e657311646e616d65722e5c7372635c6c69625c6275696c642e7273656c696e657\
    3116b746f74616c5f66696c657301";

let languages = Languages::new();
languages.get_statistics(&vec!["src/lib/build.rs"], &vec![".git"]);

assert_eq!(cbor, languages.to_cbor().to_hex());

fn to_json(&self) -> Result<String, Error>

Converts Languages to JSON.

let json = r#"{"Rust":{"blanks":5,"code":12,"comments":0,"stats":[{"blanks":5,"code":12,"comments":0,"lines":17,"name":".\\src\\lib\\build.rs"}],"lines":17,"total_files":1}}"#;
let languages = Languages::new();
languages.get_statistics(&vec!["src/lib/build.rs"], &vec![".git"]);

assert_eq!(json, languages.to_json());

fn to_yaml(&self) -> Result<String, Error>

Converts Languages to YAML.

let yaml = r#"
    ---
    "Rust":
    "blanks": 5
    "code": 12
    "comments": 0
    "lines": 17
    "stats":
        -
        "blanks": 5
        "code": 12
        "comments": 0
        "lines": 17
        "name": ".\\src\\lib\\build.rs"
    "total_files": 1"#;
let languages = Languages::new();
languages.get_statistics(&vec!["src/lib/build.rs"], &vec![".git"]);

assert_eq!(yaml, languages.to_yaml());